remove inline import, add sanitization to avoid missing ' error, move getconsole error to page utile
